🚀 Release v7.0.0 (#2502)

+++ forms-flow-web/src/components/PrivateRoute.jsx
@@ -1,9 +1,15 @@
 /* eslint-disable no-unused-vars */
-import React, { useEffect, Suspense, lazy, useMemo } from "react";
+import React, {
+  useEffect,
+  Suspense,
+  useMemo,
+  useCallback
+} from "react";
 import { Route, Switch, Redirect, useParams } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import {
   BASE_ROUTE,
+  getRoute,
   DRAFT_ENABLED,
   MULTITENANCY_ENABLED,
   KEYCLOAK_AUTH_URL,
@@ -19,9 +25,6 @@ import {
 } from "../actions/bpmActions";
 import { setLanguage } from "../actions/languageSetAction";
 import {
-  CLIENT,
-  STAFF_REVIEWER,
-  STAFF_DESIGNER,
   ENABLE_APPLICATIONS_MODULE,
   ENABLE_DASHBOARDS_MODULE,
   ENABLE_FORMS_MODULE,
@@ -31,7 +34,10 @@ import {
 
 import Loading from "../containers/Loading";
 import NotFound from "./NotFound";
-import { setTenantFromId } from "../apiManager/services/tenantServices";
+import {
+  setTenantFromId,
+  validateTenant,
+} from "../apiManager/services/tenantServices";
 
 // Lazy imports is having issues with micro-front-end build
 
@@ -51,7 +57,10 @@ import {
 import { AppConfig } from "../config";
 import { getFormioRoleIds } from "../apiManager/services/userservices";
 import AccessDenied from "./AccessDenied";
-
+import { LANGUAGE } from "../constants/constants";
+import useUserRoles from "../constants/permissions";
+import { getUserRoles } from "../apiManager/services/authorizationService"; // Assuming you have a service to get roles
+import PropTypes from "prop-types";
 export const kcServiceInstance = (tenantId = null) => {
   return KeycloakService.getInstance(
     KEYCLOAK_AUTH_URL,
@@ -75,11 +84,33 @@ const PrivateRoute = React.memo((props) => {
   const isAuth = useSelector((state) => state.user.isAuthenticated);
   const userRoles = useSelector((state) => state.user.roles || []);
   const { tenantId } = useParams();
-  const redirecUrl = MULTITENANCY_ENABLED ? `/tenant/${tenantId}/` : `/`;
   const selectedLanguage = useSelector((state) => state.user.lang);
   const tenant = useSelector((state) => state.tenants);
   const [authError, setAuthError] = React.useState(false);
   const [kcInstance, setKcInstance] = React.useState(getKcInstance());
+  const [tenantValid, setTenantValid] = React.useState(true); // State to track tenant validity
+  const ROUTE_TO = getRoute(tenantId);
+  const {
+    admin,
+    createDesigns,
+    createSubmissions,
+    viewDesigns,
+    viewSubmissions,
+    viewTasks,
+    manageTasks,
+    viewDashboards,
+  } = useUserRoles();
+
+  const BASE_ROUTE_PATH = (() => {
+    if (viewTasks || manageTasks) return ROUTE_TO.TASK;
+    if (createSubmissions) return ROUTE_TO.FORM;
+    if (createDesigns || viewDesigns) return ROUTE_TO.FORMFLOW;
+    if (admin) return ROUTE_TO.ADMIN;
+    if (viewSubmissions) return ROUTE_TO.APPLICATION;
+    if (viewDashboards) return ROUTE_TO.METRICS;
+    return ROUTE_TO.NOTFOUND;
+  })();
+
 
   const authenticate = (instance, store) => {
     setKcInstance(instance);
@@ -88,70 +119,96 @@ const PrivateRoute = React.memo((props) => {
     );
     dispatch(setUserAuth(instance.isAuthenticated()));
     store.dispatch(setUserToken(instance.getToken()));
-    //Set Cammunda/Formio Base URL
+    // Set Cammunda/Formio Base URL
     setApiBaseUrlToLocalStorage();
-    // get formio roles
+  
+    // Fetch user roles and update the local storage
+    getUserRoles()
+      .then((res) => {
+        if (res) {
+          const { data = [] } = res;
+          const roles = data.map((role) => role.name);
+          localStorage.setItem("allAvailableRoles", JSON.stringify(roles));
+        }
+      })
+      .catch((error) => console.error("Error fetching roles", error));
+  
+    // Get formio roles
     store.dispatch(
       getFormioRoleIds((err) => {
         if (err) {
           console.error(err);
-          // doLogout();
         } else {
           store.dispatch(
             setUserDetails(
               JSON.parse(StorageService.get(StorageService.User.USER_DETAILS))
             )
           );
-
-          // onAuthenticatedCallback();
         }
       })
     );
   };
+  
 
-  useEffect(() => {
+  const keycloakInitialize = useCallback(() => {
     let instance = tenantId ? kcServiceInstance(tenantId) : kcServiceInstance();
-    if (tenantId && props.store) {
-      let currentTenant = sessionStorage.getItem("tenantKey");
-      if (currentTenant && currentTenant !== tenantId) {
-        sessionStorage.clear();
-        localStorage.clear();
-      }
-      sessionStorage.setItem("tenantKey", tenantId);
-      dispatch(setTenantFromId(tenantId));
-    }
     if (props.store) {
       if (kcInstance) {
         authenticate(kcInstance, props.store);
       } else {
         instance.initKeycloak((authenticated) => {
-          if(!authenticated)
-          {
+          if (!authenticated) {
             setAuthError(true);
-          }
-          else{
+          } else {
             authenticate(instance, props.store);
             publish("FF_AUTH", instance);
           }
         });
       }
     }
-  }, [props.store, tenantId, dispatch]);
+  }, [props.store, kcInstance, tenantId]);
 
-  /**
-   * Retrieves the user's locale from the Keycloak instance or the tenant data, and dispatches an action to set the language in the application state.
-   * This effect is triggered whenever the Keycloak instance or the tenant data changes.
-   */
+
   useEffect(() => {
-    if(kcInstance){
-      const lang = kcInstance?.userData?.locale ||
-      tenant?.tenantData?.details?.locale ||
-      selectedLanguage ;
+    if (tenantId && MULTITENANCY_ENABLED) {
+      validateTenant(tenantId)
+        .then((res) => {
+          if (!res.data.tenantKeyExist) {
+            setTenantValid(false);
+          } else {
+            setTenantValid(true);
+
+            if (tenantId && props.store) {
+              let currentTenant = sessionStorage.getItem("tenantKey");
+              if (currentTenant && currentTenant !== tenantId) {
+                sessionStorage.clear();
+                localStorage.clear();
+              }
+              sessionStorage.setItem("tenantKey", tenantId);
+              dispatch(setTenantFromId(tenantId));
+              keycloakInitialize();
+            }
+          }
+        })
+        .catch((err) => {
+          console.error("Error validating tenant", err);
+          setTenantValid(false);
+        });
+    } else {
+      keycloakInitialize();
+    }
+  }, [tenantId, props.store, dispatch]);
+
+  useEffect(() => {
+    if (kcInstance) {
+      const lang =
+        kcInstance?.userData?.locale ||
+        tenant?.tenantData?.details?.locale ||
+        selectedLanguage ||
+        LANGUAGE;
       dispatch(setLanguage(lang));
     }
   }, [kcInstance, tenant?.tenantData]);
-
-  // useMemo prevents unneccessary rerendering caused by the route update.
 
   const DesignerRoute = useMemo(
     () =>
@@ -160,10 +217,28 @@ const PrivateRoute = React.memo((props) => {
           <Route
             {...rest}
             render={(props) =>
-              userRoles.includes(STAFF_DESIGNER) ? (
+              createDesigns || viewDesigns ? (
                 <Component {...props} />
               ) : (
-                <>unauthorized</>
+                <AccessDenied userRoles={userRoles} />
+              )
+            }
+          />
+        ),
+    [userRoles]
+  );
+
+  const DashBoardRoute = useMemo(
+    () =>
+      ({ component: Component, ...rest }) =>
+        (
+          <Route
+            {...rest}
+            render={(props) =>
+              viewDashboards ? (
+                <Component {...props} />
+              ) : (
+                <AccessDenied userRoles={userRoles} />
               )
             }
           />
@@ -178,10 +253,10 @@ const PrivateRoute = React.memo((props) => {
           <Route
             {...rest}
             render={(props) =>
-              userRoles.includes(STAFF_REVIEWER) ? (
+              viewTasks || manageTasks || viewDashboards ? (
                 <Component {...props} />
               ) : (
-                <>unauthorized</>
+                <AccessDenied userRoles={userRoles} />
               )
             }
           />
@@ -196,11 +271,10 @@ const PrivateRoute = React.memo((props) => {
           <Route
             {...rest}
             render={(props) =>
-              userRoles.includes(STAFF_REVIEWER) ||
-              userRoles.includes(CLIENT) ? (
+              viewSubmissions ? (
                 <Component {...props} />
               ) : (
-                <>unauthorized</>
+                <AccessDenied userRoles={userRoles} />
               )
             }
           />
@@ -215,78 +289,101 @@ const PrivateRoute = React.memo((props) => {
           <Route
             {...rest}
             render={(props) =>
-              DRAFT_ENABLED &&
-              (userRoles.includes(STAFF_REVIEWER) ||
-                userRoles.includes(CLIENT)) ? (
+              DRAFT_ENABLED && viewSubmissions ? (
                 <Component {...props} />
               ) : (
-                <>unauthorized</>
+                <AccessDenied userRoles={userRoles} />
               )
             }
           />
         ),
     [userRoles]
   );
+
+  const ClientRoute = useMemo(
+    () =>
+      ({ component: Component, ...rest }) =>
+      (
+        <Route
+          {...rest}
+          render={(props) =>
+            createSubmissions || viewSubmissions  ? (
+              <Component {...props} />
+            ) : (
+              <AccessDenied userRoles={userRoles} />
+            )
+          }
+        />
+      ),
+    [userRoles]
+  );
+
+  ClientRoute.propTypes = {
+    component: PropTypes.elementType.isRequired,
+  };
+
+  if (!tenantValid) {
+    return <NotFound />;
+  }
+
   return (
     <>
       {authError ? (
-        <AccessDenied />
+        <AccessDenied userRoles={userRoles} />
       ) : isAuth ? (
         <Suspense fallback={<Loading />}>
           <Switch>
             {ENABLE_FORMS_MODULE && (
-              <Route path={`${BASE_ROUTE}form`} component={Form} />
+              <ClientRoute path={ROUTE_TO.FORM} component={Form} />
             )}
             {ENABLE_FORMS_MODULE && (
-              <DesignerRoute path={`${BASE_ROUTE}formflow`} component={Form} />
+              <DesignerRoute path={ROUTE_TO.FORMFLOW} component={Form} />
             )}
             {ENABLE_APPLICATIONS_MODULE && (
-              <DraftRoute path={`${BASE_ROUTE}draft`} component={Drafts} />
+              <DraftRoute path={ROUTE_TO.DRAFT} component={Drafts} />
             )}
             {ENABLE_APPLICATIONS_MODULE && (
               <ClientReviewerRoute
-                path={`${BASE_ROUTE}application`}
+                path={ROUTE_TO.APPLICATION}
                 component={Application}
               />
             )}
-
             {ENABLE_PROCESSES_MODULE && (
               <DesignerRoute
-                path={`${BASE_ROUTE}processes`}
+                path={ROUTE_TO.SUBFLOW}
                 component={Modeler}
               />
             )}
-
+            {ENABLE_PROCESSES_MODULE && (
+              <DesignerRoute
+                path={ROUTE_TO.DECISIONTABLE}
+                component={Modeler}
+              />
+            )}
             {ENABLE_DASHBOARDS_MODULE && (
-              <ReviewerRoute
-                path={`${BASE_ROUTE}metrics`}
+              <DashBoardRoute
+                path={ROUTE_TO.METRICS}
                 component={DashboardPage}
               />
             )}
             {ENABLE_DASHBOARDS_MODULE && (
-              <ReviewerRoute
-                path={`${BASE_ROUTE}insights`}
+              <DashBoardRoute
+                path={ROUTE_TO.INSIGHTS}
                 component={InsightsPage}
               />
             )}
             {ENABLE_TASKS_MODULE && (
               <ReviewerRoute
-                path={`${BASE_ROUTE}task`}
+                path={ROUTE_TO.TASK}
                 component={ServiceFlow}
               />
             )}
-
+            <Route exact path={ROUTE_TO.ADMIN} /> 
             <Route exact path={BASE_ROUTE}>
-              {userRoles.length && <Redirect
-                to={
-                  userRoles?.includes(STAFF_REVIEWER)
-                    ? `${redirecUrl}task`
-                    : `${redirecUrl}form`
-                }
-              />}
+              {userRoles.length && <Redirect to={BASE_ROUTE_PATH} />}
             </Route>
-            <Route path="/404" exact={true} component={NotFound} />
-            <Redirect from="*" to="/404" />
+            <Route path={ROUTE_TO.NOTFOUND} exact={true} component={NotFound} />
+            <Redirect from="*" to={ROUTE_TO.NOTFOUND} />
           </Switch>
         </Suspense>
       ) : (
